+++ aqueduct/lib/src/application/isolate_application_server.dart
@@ -66,7 +66,7 @@ class ApplicationIsolateServer extends ApplicationServer {
     logger.fine("ApplicationIsolateServer($identifier) closing server");
     await close();
     logger.fine("ApplicationIsolateServer($identifier) did close server");
-    await ApplicationServiceRegistry.defaultInstance.close();
+    await ServiceRegistry.defaultInstance.close();
     logger.clearListeners();
     logger.fine(
         "ApplicationIsolateServer($identifier) sending stop acknowledgement");