+++ aqueduct/lib/src/http/serializable.dart
@@ -8,19 +8,19 @@ import 'http.dart';
 /// Interface for serializable instances to be decoded from an HTTP request body and encoded to an HTTP response body.
 ///
 /// Implementers of this interface may be a [Response.body] and bound with an [Bind.body] in [ResourceController].
-abstract class HTTPSerializable {
+abstract class Serializable {
   /// Documents [serializable].
   ///
-  /// [serializable] must implement, extend or mixin [HTTPSerializable]. The returned [APISchemaObject]
+  /// [serializable] must implement, extend or mixin [Serializable]. The returned [APISchemaObject]
   /// will be of type 'object'. Each instance variable declared in [serializable] will be a property of this object.
   /// Instance variables are documented according to [APIComponentDocumenter.documentVariable]. See the API reference
   /// for this method for supported types.
   static APISchemaObject document(
       APIDocumentContext context, Type serializable) {
     final mirror = reflectClass(serializable);
-    if (!mirror.isAssignableTo(reflectType(HTTPSerializable))) {
+    if (!mirror.isAssignableTo(reflectType(Serializable))) {
       throw ArgumentError(
-          "Cannot document '${MirrorSystem.getName(mirror.simpleName)}' as 'HTTPSerializable', because it is not an 'HTTPSerializable'.");
+          "Cannot document '${MirrorSystem.getName(mirror.simpleName)}' as 'Serializable', because it is not an 'Serializable'.");
     }
 
     final properties = <String, APISchemaObject>{};
@@ -51,7 +51,7 @@ abstract class HTTPSerializable {
   ///
   /// This method returns a [Map<String, dynamic>] where each key is the name of a property in the implementing type.
   /// If a [Response.body]'s type implements this interface, this method is invoked prior to any content-type encoding
-  /// performed by the [Response].  A [Response.body] may also be a [List<HTTPSerializable>], for which this method is invoked on
+  /// performed by the [Response].  A [Response.body] may also be a [List<Serializable>], for which this method is invoked on
   /// each element in the list.
   Map<String, dynamic> asMap();
 }